Handle data type and JSON field name case-insensivity.
In the state tracker in the server and in the simulator we need to
use the exact same case-sensitivity checking type as in the Sensorino
firmware, and this was a little tricky.  Solve it everywhere for once.

Also add more warnings on unknown datatypes.

Fix typos in service description parsing, now tested.