Handle node addrs transparently in client side, change notification order.
Drop the forcing of node addresses in the client side code to be an int,
it can now also be a string (e.g. an IP) and the client-side should just
always use the same type as it received.

Also only call notifiers on state changes after updating the state, so
that the notifiers can safely query new state (if needed.. shouldn't be
needed anyway).