Handle messages from remote rule engines.
For now, we handle incoming SET and REQUEST messages exactly the
same as those generated by ourselves.  In the database only the
console log will show information about who the message came from.
The state tables show no difference at this time.  We queue up
the SET and REQUEST incoming messages so we can keep track of
their error status, just the same as for outgoing.

Since the rule engine will cause a SET message to be generated
by the remote end immediately after the PUBLISH that triggered the
rule, we get a sequence of messages that we should no interrupt.
Since there's no flow control, we really should start queuing our
outgoing messages, such as the discovery agent ones, with delays
so the outgoing message triggered on our side doesn't clash with
one triggered on the remote side.