By default notify general listeners of important changes only.
When they subscribe they have to pass true as the second argument to
get the old behaviour where they'll get notified of every change.  The
unimportant changes are the ones to channel values, all other changes
"important" or "meta".

This is to reduce the number of times the browser widget or the
floorplan unused channel check need to recalculate their stuff since
they're only interested in the "meta" changes.